moves histogram code out of profiler
