+++ testbenches/common/v/vanilla_core_pc_histogram.def
@@ -1,0 +1,1 @@
+`DEFINE_PC_STATE()
